requirements.txt
torch
diffusers
transformers
accelerate
safetensors
pillow
numpy
rembg[cpu]
opencv-python